Condense CivSet view into base class
