+++ app/grandchallenge/reader_studies/views.py
@@ -60,6 +60,7 @@ from grandchallenge.components.serializers import (
 from grandchallenge.components.views import (
     CIVSetBulkDelete,
     CIVSetDelete,
+    CIVSetDetail,
     CIVSetFormMixin,
     CivSetListView,
     FileUpdateBaseView,
@@ -1261,24 +1262,15 @@ class QuestionWidgetsView(BaseAddObjectToReaderStudyMixin, View):
         return HttpResponse(form["widget"])
 
 
-class DisplaySetDetailView(
-    LoginRequiredMixin, ObjectPermissionRequiredMixin, DetailView
-):
+class DisplaySetDetailView(CIVSetDetail):
     model = DisplaySet
-    template_name = "components/civ_set_detail.html"
-
     permission_required = (
         f"{ReaderStudy._meta.app_label}.view_{DisplaySet._meta.model_name}"
     )
-    raise_exception = True
 
-    def get_context_data(self, *args, **kwargs):
-        context = super().get_context_data(*args, **kwargs)
-
-        object = self.get_object()
-        context.update(dict(base_object=object.base_object))
-
-        return context
+    @property
+    def base_object(self):
+        return self.object.base_object
 
 
 class DisplaySetUpdateView(
